Cleanup startup process
